chore(ios/swift): move HeaderEndActions definition into Header
