+++ ios/ChatExample/ChatExample/SDK/Header.swift
@@ -198,3 +198,33 @@ internal func convertToUIView<V: View>(_ swiftUIView: V) -> UIView {
     controller.view.backgroundColor = .clear
     return view
 }
+
+///
+/// End actions to be used in a `Header` via `Header#withEndAction`
+///
+public enum EndAction {
+    /// Renders an End Action with a visible 'Next' label
+    ///
+    /// - Parameters:
+    ///  - onClick: Callback for when the action is clicked.
+    ///
+    case next(onClick: () -> Void)
+    /// Renders an End Action with a visible 'Save' label
+    ///
+    /// - Parameters:
+    ///  - onClick: Callback for when the action is clicked.
+    ///
+    case save(onClick: () -> Void)
+    /// Renders an End Action with a visible 'Menu' label
+    ///
+    /// - Parameters:
+    ///  - onClick: Callback for when the action is clicked.
+    ///
+    case menu(onClick: () -> Void)
+    /// Renders an End Action with a visible 'Create' label
+    ///
+    /// - Parameters:
+    ///  - onClick: Callback for when the action is clicked.
+    ///
+    case create(onClick: () -> Void)
+}
